requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching
redis==5.0.1
hiredis==2.3.2

# ML & NLP
torch==2.1.2
transformers==4.37.2
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3

# Vector DB
pinecone-client==3.0.2

# Observability
prometheus-client==0.19.0
structlog==24.1.0

# Circuit breaker & resilience
tenacity==8.2.3
pybreaker==1.0.1

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
